RIP-354 Dissalow multiple logins
https://eaflood.atlassian.net/browse/RIP-354

Adds the devise security extensions gem which adds support for
limiting concurrent logins.

To test, sign in as a user in Chrome. Open Firefox and sign in as the
same user. Then back in Chrome try to use the application. You should
be redirected to a page with an alert informing you someone else is
using the account.