Remove dependency on before_commit for CI checks
[before_commit](https://github.com/EnvironmentAgency/before_commit) was an attempt to have a central gem which could be included in every Ruby based project. The gem would manage the checks we wished to perform in each, thereby ensuring consistency and quality across the projects.

In practise we have found when it works, it works great however when it doesn't, it can be very confusing to try and resolve what is actually failing (the project, the check, or **before_commit** itself). We also frequently encountered issues with trying to essentially enforce a centrally controlled workflow on differing development environments.

We have agreed instead to move to ensuring the checks are run independently, and only on the relevant CI server. It will be the responsibility of devs to run these same checks locally, in the way appropriate to their environment, to ensure they don't '**break the build**'.

This change covers removing the dependency on **before_commit** and instead moving to the new agreed model. This involves

- removing **before_commit** references
- removing configuration specific to **before_commit**
- removing usage of **before_commit** from the `.circle.yml` file
- adding back each of the checks **before_commit** was handling, via gems, config files, updates to `.circle.yml` or combinations of them
- set default task in Rakefile to run **rspec**
- add task to Rakefile to run rubocop
- adding our standard `.rubocop.yml` to the project
- removing matchers in `.gitignore` which would prevent files like `.rubocop.yml` from being committed
- resolving any **Rubocop** issues found
- resolving any **i18n-tasks** 'unused' errors found
- add missing license (just cause its private doesn't mean it shouldn't have a license)
- updating documentation (README.md) to reflect the changes made