Remove overcommit from repo
When FRAE was being built the team experimented with using [overcommit](https://github.com/sds/overcommit) to ensure quality standards were maintained on each commit.

In practise they found it got in the way of perfectly valid commits, and ran into problems trying to deal with each dev's preferred setup.

Hence we stopped using it and stripped it from Waste Exemptions. Now we're doing the same for FRAE 😀