Remove heroku from repo (#98)
When the app was first built there were no environments in which to deploy it. So the team used Heroku whilst they waited for them to become available.

The heroku apps are no longer used and hence there is no longer a need to maintain anything related to Heroku in the project.