Move github changelog generator to main group (#124)
The original code for the github changelog generator entry came from [Waste Carriers](https://github.com/DEFRA/waste-carriers-back-office). In the original entry its not in the `development` group but in the main group. A note explains this should be in the `development` group, but due to an error caused by the gem when deplying to heroku, it was moved to the main group.

Well, we're deploying to heroku anymore so PR #227 added it to the `development` group!

🤦 Well guess what happened when we then deployed this to our environments!?

So it seems the original comment was half right. Though it should be in the development group, putting the entry their breaks any deployment.