Migrate area lookup to defra-ruby-area (#144)
https://eaflood.atlassian.net/browse/RUBY-484

We have switched the FRAE service to use the gem [defra-ruby-area](https://github.com/DEFRA/defra-ruby-area) instead of [ea-area_lookup](https://github.com/DEFRA/ea-area_lookup). This means we need to update the engine (where the gem is added as a dependency and the logic sits) and remove the old gem's initializer (the new gem doesn't need one).