Add scheduled task to add missing Area information
From: https://eaflood.atlassian.net/browse/RUBY-604

Add a scheduled rake task that will run the Area lookup job for all locations that are missing the info.
The task contains a time constraint that will exit the job after it has run for X time. The X time is ENV configurable and default to 60 minutes.