Remove some of the black list keys
In our recent work we do not believe we need to go to this level to strip values from the errors we send to Errbit.

So this change removes some, which are either transient i.e. they are just related to temporary data, or which are likely to appear on the public register for flood risk exemptions.