Alter airbrake config to allow use in development (#170)
The previous team took the decision to not activate airbrake in non-production environments, no doubt to reduce noise.

We on the other hand value having it enabled in all environments as it might highlight issues we haven't spotted. It also reduces the differences between the app running in production, and running in other environments.

On that last point, PR #152 added a rake task which when run will update the area for enrollments missing it. We thought all was fine until we tried to run it in an environment where `RAILS_ENV=production`.

We get an error caused by the `Airbrake.close` line, something we would not have come across because Airbrake is not enabled when running locally.

So, for this reason, we're making these changes, which bring the config more inline with our other services, and ensure Airbrake is always enabled by default irrespective of environment.

## Remove some of the black-list keys

In our recent work, we do not believe we need to go to this level to strip values from the errors we send to Errbit.

So this change removes some, which are either transient i.e. they are just related to temporary data, or which are likely to appear on the public register for flood risk exemptions.