Update production user seeds with test emails
In our [WEX](https://github.com/DEFRA/ruby-services-team/tree/master/services/wex) and [WCR](https://github.com/DEFRA/ruby-services-team/tree/master/services/wcr) services we have a mechanism for seeding both locally and in our non-production AWS environments. This is a big help when it comes to testing as it allows us to reset the database, re-seed, and start with a clean slate.

Unfortunately the FRAE service does not have the same setup. It relies on environment specific seed files. This is great for development, but it protects production by having nothing. In actual production there was a one off rake task that was used to seed the database. It is based on `production_users.yml` that as a list of users needed at the time the service went live 3 years ago.

Rather than mess with the current seeding mechanism (it goes quite different in the engine) this makes the following changes

- it removes `production_users.yml` and the associate rake task
- it moves the code in `development.rb` into `all.rb`
- it adds an env var flag that must be setting to allow seeding to take place