Update seeding match WEX
This change removes the original seeding methodology and uses the simplier version from WEX.

With only `all.rb` and `development.rb` still being the only seed files after 3 years of existing its clear we don't need the flexibility the previous design had.

So this change switches over way the data is seeded to match (almost) the way WEX works.

A key difference is we can't use the `find_or_create()` method because FRAE uses [Rolify](https://github.com/RolifyCommunity/rolify) which means roles are handled quite differently. For instance roles are created via the user instance, but then added to the relevant user in an after hook. So we appear to need a persisted user instance before we can apply a role to it.

Hence the new check to see if the user already exists, and if not simply create it.