Fix default whenever log path
https://eaflood.atlassian.net/browse/RUBY-604

When trying to test the job in our AWS environments it didn't run. After some investigation spotted that its because the default log path does not actually match our environment.

For 'reasons' even though the project is called `flood-risk-back-office` the folder created in the environment is  `flood-risk-activity-exemption-back-office`. Hence the path did not exist so the cron job would fail when ran.

This fix should sort the problem.