Mount defra ruby email engine directly 😩
Sooo, just accessing the defra-ruby-email engine via the flood-risk-engine isn't going to work.

We assumed based on the behaviour of the `/pages` routes that it would. Having mounted the defra-ruby-email in the engine we'd be able to access it at `/email/last-email` without the need to authenticate. Turns out we were wrong on both counts.

- you have to go to `/fre/email/last-email`
- you have to be logged in first else you just get redirected to the sign-in page

We believe that something in the HighVoltage gem means it is always handling any requests to `/pages` directly. And as the gem comes in as a dependency of the flood-risk-engine, a request to `/pages/version` just works. If you check the engine's routes file you'll see that HighVoltage is not mounted, so its doing its own magic behind the scenes.

FRAE works differently to WCR and WEX because it wraps the mounting of the engine in a Devise `authenticate` block and puts it on a path other than `/`.

```ruby
authenticate :user do
  mount FloodRiskEngine::Engine => "/fre"
end
```

So as defra-ruby-email doesn't have the same magic as HighVoltage

- we have to access it through `/fre`
- we have to be authenticated before we can access anything in `/fre`

So other than deep diving on HighVoltage and re-writing defra-ruby-email, the only solution is to setup the back office to use the engine directly. This change does that.