Enable defra-ruby-email in the project (#209)
Currently the [Waste Exemptions service](https://github.com/DEFRA/waste-exemptions-engine/pull/112) and the [Waste Carriers Frontend](https://github.com/DEFRA/waste-carriers-frontend/pull/227) have the ability to intercept and playback the details of the last email sent.

This feature is only enabled in our non-production environments and is used as part of our acceptance tests. However, our QA @andrewhick has rightly pointed out that the functionality is inconsistent across our services.

Waste Exemptions has it throughout, Waste Carriers only has it in the old app, and Flood Risk Activity Exemptions doesn't have it at all. This makes writing and maintaining acceptance tests across the 3 services difficult and inconsistent so we have been asked to resolve the issue.

Rather than duplicate the existing code even more, we have created [defra-ruby-email](https://github.com/DEFRA/defra-ruby-email) a reusable engine that can be mounted into an application to provide the same functionality.

Having created the gem, we implemented it into the [flood-risk-engine](https://github.com/DEFRA/flood-risk-engine/pull/332) to add the feature to this service.

This change completes the implementation for the back-office.

##  Need to mount defra-ruby-email engine directly

Sooo, just accessing the defra-ruby-email engine via the flood-risk-engine doesn't work 😩.

We assumed based on the behaviour of the `/pages` routes that it would. Having mounted the defra-ruby-email in the engine we'd be able to access it at `/email/last-email` without the need to authenticate. Turns out we were wrong on both counts.

- you have to go to `/fre/email/last-email`
- you have to be logged in first else you just get redirected to the sign-in page

We believe that something in the HighVoltage gem means it is always handling any requests to `/pages` directly. And as the gem comes in as a dependency of the flood-risk-engine, a request to `/pages/version` just works. If you check the engine's routes file you'll see that HighVoltage is not mounted, so it's doing its own magic behind the scenes.

FRAE works differently to WCR and WEX because it wraps the mounting of the engine in a Devise `authenticate` block and puts it on a path other than `/`.

```ruby
authenticate :user do
  mount FloodRiskEngine::Engine => "/fre"
end
```

So as defra-ruby-email doesn't have the same magic as HighVoltage

- we have to access it through `/fre`
- we have to be authenticated before we can access anything in `/fre`

So other than deep diving on HighVoltage and re-writing defra-ruby-email, the only solution was to set up the back office to use the engine directly.