Update project to support AWS S3 encryption (#236)
https://eaflood.atlassian.net/browse/RUBY-1099

As part of web-ops ongoing work to in support of their AWS security review, they are ensuring all environments and resources are encrypted as by default.

This includes AWS S3 buckets. Currently, they have enabled the feature in the FRAE AWS DEV environment. This change covers everything needed to get the app working in that environment.