Update defra-ruby-aws to encryption configurable (#242)
https://eaflood.atlassian.net/browse/RUBY-1156

In PR #238 recently we updated [defra-ruby-aws](https://github.com/DEFRA/defra-ruby-aws) to a version that encrypted our uploads using AWS:KMS instead of AES256.

It turns out that our 3rd parties are unable to decrypt those files using this new encryption method. To allow them to is going to take some reconfiguration of our S3 buckets and that will impact a number of users. So we need to revert the change.

However, rather than going back to the previous version of the gem, we have updated it to make switching between encryption methods configurable. This updates the back-office to use the updated gem and prepare it for being able to switch using config in the future.