Validate the postcode length
Relates to:
https://eaflood.atlassian.net/browse/RUBY-1495?focusedCommentId=332672

There is a database constraint on the postcode column - max 8 chars length
Previously we used the postcode validator to check the format of the postcode.
This opened a can of worms because it also does a postcode look up.
As this form is to be used by NCCC to fix incorrect addresses, we are
going with a manual version.