Update engine without legacy code
https://eaflood.atlassian.net/browse/RUBY-1592

There are some components (notably Reform) that have been removed from the engine but are still being refactored out of the back office. So I've added some things back in temporarily.

Also noticed that we'll need to change the way we store the AD status in the engine - this will be done in an upcoming PR.